+++ src/main/java/janggi/domain/board/Board.java
@@ -6,7 +6,7 @@ import janggi.domain.piece.Empty;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.PieceInfo;
 import janggi.domain.piece.PieceType;
-import janggi.domain.piece.Route;
+import janggi.domain.Route;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
