+++ src/main/java/janggi/domain/piece/LinearPiece.java
@@ -2,6 +2,7 @@ package janggi.domain.piece;
 
 import janggi.domain.Movement;
 import janggi.domain.Position;
+import janggi.domain.Route;
 import janggi.domain.Side;
 import janggi.domain.board.BoardInterface;
 import janggi.domain.policy.RoutePolicy;
