+++ src/main/java/janggi/domain/piece/Piece.java
@@ -1,6 +1,7 @@
 package janggi.domain.piece;
 
 import janggi.domain.Position;
+import janggi.domain.Route;
 import janggi.domain.Side;
 import janggi.domain.board.BoardInterface;
