+++ src/main/java/janggi/domain/policy/ClearPathPolicy.java
@@ -2,7 +2,7 @@ package janggi.domain.policy;
 
 import janggi.domain.Side;
 import janggi.domain.board.BoardInterface;
-import janggi.domain.piece.Route;
+import janggi.domain.Route;
 
 public class ClearPathPolicy implements RoutePolicy {
     @Override
