+++ src/test/java/janggi/domain/piece/PoTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import janggi.domain.Position;
+import janggi.domain.Route;
 import janggi.domain.Side;
 import java.util.List;
 import org.junit.jupiter.api.Test;
