+++ src/test/java/janggi/domain/policy/ClearPathPolicyTest.java
@@ -6,7 +6,7 @@ import janggi.domain.Position;
 import janggi.domain.Side;
 import janggi.domain.board.BoardInterface;
 import janggi.domain.piece.PieceType;
-import janggi.domain.piece.Route;
+import janggi.domain.Route;
 import java.util.List;
 import org.junit.jupiter.api.Test;
